refactor(workspace): :truck: moved `testUtils.tsx` into the test utils folder
